Handle comments and other 'objects' for Bugzilla 5+ messages
This was completely broken. In the old days, when we queried
Bugzilla for the bug data, we got a 'comments' dict containing
all(?) the comments on the bug, and then if we decided this was
a 'added a comment' type message, we tried to find the 'correct'
comment from all the comments and include that in the top-level
message data.

Now, though, we aren't querying Bugzilla, we're assuming we get
all the right info from the message Bugzilla sends us. And we
are, but not at all in the format we were expecting.

When a comment is created, the message Bugzilla emits is rather
different from the message for bug.create or bug.modify. It does
not have a top-level 'bug' key because the 'object' of the msg
is not considered to be the bug, but the comment. So it has a
top-level 'comment' key with the data about the comment, and
*inside* that 'comment' dict is the 'bug' dict with all the
generic data about the bug.

This is actually a generic mechanism: 'comment' isn't the only
thing it applies to. It also applies to (at least) 'attachment':
there are 'attachment.new' messages when an attachment is
created, with an 'attachment' dict in them. So we should handle
this in a generic way.

We throw out the expected 'comments' field from the bug dict
data (it seems BZ never sends us this, even in bug.modify
messages), and we also throw out `find_relevant_item` since we
no longer need to do that stuff. We handle messages where the
object dict is not 'bug' and unpick the bug dict from the object
dict, and include the object data into the fedmsg we emit
properly. We also stick an 'author' into comment dicts for
fedmsg.meta compatibility.

Signed-off-by: Adam Williamson <awilliam@redhat.com>