Fix decision as to whether this is a 'new bug' event
This is again wrong for BZ 5.0 messages...as we can see, since a
bug.new event has not been emitted since 2018-12-08. In fact,
we're lucky this didn't blow up - it assumes bug['comments']
always exists, but it doesn't any more. We only don't crash on
that because the *first* part of the conditional always fails;
with BZ 5.0, *all* messages include an 'event'.

This should be the correct way to make the decision now - just
by referring to the upstream message topic.

Signed-off-by: Adam Williamson <awilliam@redhat.com>