Add a test suite
For now, covers bug modification, comment creation and dropping
of private and off-topic messages. Doesn't yet cover new bug
creation or datetime conversion because both of those things
are broken (I will fix this and add tests in later commits).

Also adjusts the tox config to run the tests properly, fixes or
silences a few errors in unrelated files so the flake8 test
passes, and extends the tox environments for newer Python 3
versions.

Signed-off-by: Adam Williamson <awilliam@redhat.com>