Fix convert_datetimes for Bugzilla 5.0 messages
The existing implementation is still expecting to work with
data from querying Bugzilla directly, as we used to do - it
looks for Python objects with a particular attribute. Now we're
not querying Bugzilla ourselves but just working with the
upstream message, that doesn't work at all any more, so it was
just passing everything through. This adapts it to identify and
convert the ISO-8601ish strings we now get in the upstream
messages.

Signed-off-by: Adam Williamson <awilliam@redhat.com>