Handle 'reporter' vs. 'creator' with backwards compat
The upstream messages don't have 'creator' as the reporter's
email address, they have 'reporter' as a dict with various
values. Include the raw upstream 'reporter' dict and create
a backwards-compatible 'author' from it.

I would like to fix the meta processor and drop all of this
backwards compat stuff, but until we can get that done, let's
make it as damn backwards-compatible as humanly possible!

Signed-off-by: Adam Williamson <awilliam@redhat.com>