Don't 'organize' the message as it comes off the BZ wire
This is a throwback to the pre-Bugzilla 5 system where we were
querying Bugzilla directly and transforming the query result
into a message. The main idea, I think, was to strip out some
bits of the query result that were irrelevant or too heavy to
include in the message.

Now what we're operating on is already a message bus message,
I think we should trust that upstream formats it appropriately
and drop this kind of tinkering. We really should just pass on
the upstream message with as little fiddling as possible.

Signed-off-by: Adam Williamson <awilliam@redhat.com>