Extend backwards compatibility a bit
Looking at the fedmsg_meta parser for Bugzilla messages, I saw
two expectations that the current 'backwards compatibility'
doesn't meet: the parser expects a 'weburl' in the bug dict,
and it expects the 'product' to be a string not a dict (much
like 'component'). This adds those to the 'backwards compat'
block, and fixes up some of the conditionals in there a bit too.

Signed-off-by: Adam Williamson <awilliam@redhat.com>