Make Bugzilla 4 backwards compatibility transformation optional
I think this 'backwards compatibility' bodge is problematic, and
we should aim to remove it after ensuring all consumers can
handle messages without it. The bits which just add properties
that used to be in the old messages aren't so bad, but the bits
that destructively remove information (replacing dicts with
single strings) are obviously bad, and really I think this tool
should do as little as possible - the messages we relay should
look as close to the upstream messages as is feasible.

So, this makes the compatibility changes optional: you can turn
them on or off in the config (they default to 'on' for now).
This way we can declare both "regular" and "bz4 compat" schemas
(in future) and we can go look for things that consume Bugzilla
messages and make sure they handle non-'compat' messages, then
we can just turn the compat off in config without needing to
change this code again. Then some time in a later release we can
drop the compat code entirely.

Signed-off-by: Adam Williamson <awilliam@redhat.com>