Implement message schemas
This is a big fedora-messaging thing that wasn't done in the
initial port: we are supposed to define and use message schemas.
This defines two schemas, one for the current message format
*without* BZ4 backwards compatibility and one for the current
message format *with* it. It implements most of the standard
fedora-messaging methods, plus some additional convenience ones.
They handle the differences between the backwards-compatible
format and the non-backwards-compatible format as smoothly as
possible.

This was based on the bones of the fedmsg_meta processor, but
obviously heavily modified.

Note there's one fairly significant thing missing here: full
capability for working out FAS user names from Bugzilla email
addresses. This is something that fedmsg_meta provided for us in
a module called fasshim; fedora-messaging does not appear to have
an equivalent yet. Places where we need this are marked FIXME.

Signed-off-by: Adam Williamson <awilliam@redhat.com>